delete plant option on details page added + sucess message.
