+++ pages/_app.js
@@ -1,5 +1,7 @@
-import useSWR, { SWRConfig } from "swr";
+import { mutate, SWRConfig } from "swr";
 import GlobalStyle from "../styles";
+import { useRouter } from "next/router";
+import toast from "react-hot-toast";
 
 const fetcher = async (url) => {
   const res = await fetch(url);
@@ -16,11 +18,25 @@ const fetcher = async (url) => {
 };
 
 export default function App({ Component, pageProps }) {
+  const router = useRouter();
+
+  async function handleDeletePlant(id) {
+    const response = await fetch(`/api/plants/${id}`, {
+      method: "DELETE",
+    });
+
+    if (response.ok) {
+      toast.success("Plant sucessfully deleted!");
+      await mutate("/api/plants");
+      router.push("/");
+    }
+  }
+
   return (
     <>
       <GlobalStyle />
       <SWRConfig value={{ fetcher }}>
-        <Component {...pageProps} plants={plants} mutate={mutate} />
+        <Component {...pageProps} handleDeletePlant={handleDeletePlant} />
       </SWRConfig>
     </>
   );
